Updated login methods
Removed the playerfile line PlayerNeedsUpdating and added in a new one,
JustLoggedIn.
Every time someone dies or spawns in (Times when the playerfile needs
updating) the log_initstring function is called. This has therefore made
PlayerNeedsUpdating redundant, and I have moved the player updating to
that function. The field JustLoggedIn makes it so that the player is
only given the login message when they actually log in, rather than
every time their inventory is initialized. This caused a few problems
with /player_info not returning the expected data, but adding a 0.5 s
sleep and fiddling with some of the lines in log_playerinfo fixed this.