Custom spawning
Added in sector change based sector spawning. Whenever a player changes
a sector, their heat is incremented (and then decremented after 180s)
and that determines a chance of enemies spawning. If they spawn, a
random amount of ships spawn and a set BP is chosen for each ship.
Limits and config options have all been added