Fixes & templates
Fixed a bug I found with the boarding system, and added in a template
system for playerfiles and factionfiles. Now, create_playerfile and
create_factionfile make the files via the templates (located inside the
templates function to keep things neat within np++)
The update_playerfile and update_factionfile functions use the templates
to ensure the playerfiles and faction files that exist are complete.