Sector Income and FacMail
Added in multiple features this time around; The ability to receive
faction mail (similar to the player mailbox) sent by server services to
the faction
Sectors now generate income based on the number of surrounding sectors
daily, but increases per sector diminishes (0=100% base income 1=>200%
base income .... 6=<600% base income). There is also a limit to the
number of credits a beacon can hold. When full, the Beacon sends a mail
to the faction
Sectors lose their claims if the sectoral claim beacon is destroyed
aswell as sending a mail message to the faction informing about the loss
of the beacon
Buging a sector also sends a mail to the faction, informing about the
sucessfull purchase